requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0
scikit-learn>=1.3.0
joblib>=1.3.0
sqlalchemy>=2.0.0
httpx>=0.24.0
python-dotenv>=1.0.0
python-multipart>=0.0.6